Better on-disk format for JS libraries
Currently, we store `js_library` artifacts on disk a .tgz files of the sources
of just that target and `js_binary` artifacts as a .tgz containing all of that
target's transitive dependencies.

These .tgz files are created using Bazel's built-in `pkg_tar` rules. When
combining .tgz files together, each is opened serially, and each file is
extracted into the new archive. It is a very resource-intensive operation.

In general, creating the library archives isn't particularly slow. It's combining
many libraries into a binary target (such as every test).

This commit adds the stupidest on-disk format goin'. It's a gzip stream
contining a json object with the file's name and size, then its contents. Since
the goal is fast concatenation, it should be plain to see that combining
`a.jsar` and `b.jsar` is as fast/simple as `cat a.jsar b.jsar > c.jsar`.

The `js_bundle` rules really reached deep in to the `rules_js` implementaion, so
they've been rewritten at a higher level.

Change-Id: Ie20d1e9f4e00f19a39c55fc579dc7d2818a4eb67